+++ test/partitioning.cpp
@@ -27,16 +27,14 @@ void test_partitioning() {
         BULK_SKIP_SECTION_IF("Partitionings", p <= 1);
 
         BULK_SECTION("Cyclic partitioning") {
-            auto part =
-                bulk::cyclic_partitioning<2, 2>({10 * N, 10 * N}, {N, N});
+            auto part = bulk::cyclic_partitioning<2, 2>({10 * N, 10 * N}, {N, N});
             BULK_CHECK(part.multi_owner({4, 3})[0] == 4 % N,
                        "compute correctly the cyclic owner");
             BULK_CHECK(part.local_size({s % N, s / N})[0] == 10,
                        "compute correctly the cyclic size");
             BULK_CHECK(part.local({4, 3})[0] == 4 / N,
                        "compute correctly the cyclic local index");
-            BULK_CHECK(part.global({1, 1}, {s % N, s / N})[0] ==
-                           N + (s % N),
+            BULK_CHECK(part.global({1, 1}, {s % N, s / N})[0] == N + (s % N),
                        "compute correctly the cyclic global index");
         }
 
@@ -53,8 +51,7 @@ void test_partitioning() {
         }
 
         BULK_SECTION("Block partitioning") {
-            auto part =
-                bulk::block_partitioning<2, 2>({10 * N, 10 * N}, {N, N});
+            auto part = bulk::block_partitioning<2, 2>({10 * N, 10 * N}, {N, N});
             BULK_CHECK(part.multi_owner({2 * 10 + 3, 3})[0] == 2,
                        "compute correctly the block owner");
             BULK_CHECK(part.local_size({s % N, s / N})[0] == 10,
@@ -97,16 +94,15 @@ void test_partitioning() {
 
         BULK_SECTION("Binary-split-partitioning") {
             using dir = bulk::util::binary_tree<bulk::util::split>::dir;
-            auto tree = bulk::util::binary_tree<bulk::util::split>(
-                bulk::util::split{0, 4});
+            auto tree =
+            bulk::util::binary_tree<bulk::util::split>(bulk::util::split{0, 4});
             auto root = tree.root.get();
             tree.add(root, dir::left, bulk::util::split{1, 4});
             tree.add(root, dir::right, bulk::util::split{1, 4});
 
-            auto part =
-                bulk::tree_partitioning<2>({10, 10}, 4, std::move(tree));
+            auto part = bulk::tree_partitioning<2>({10, 10}, 4, std::move(tree));
 
-            BULK_CHECK((part.local_size(0) == bulk::index<2>{5, 5}),
+            BULK_CHECK((part.local_size(1) == bulk::index<2>{5, 5}),
                        "extent of bspart are correct");
 
             BULK_CHECK((part.owner({1, 1}) == part.owner({2, 2})),
@@ -123,7 +119,7 @@ void test_partitioning() {
                        "assign correct origin bspart (3)");
 
             BULK_CHECK((part.local({6, 6}) == bulk::index<2>{1, 1}),
-                       "assign correct origin bspart");
+                       "local computation bspart");
         }
 
         BULK_SECTION("Partitioned array") {
@@ -149,16 +145,32 @@ void test_partitioning() {
         }
 
         BULK_SECTION("Irregular block partitioning") {
-          auto part = bulk::block_partitioning<1>(5, 4);
-          BULK_CHECK(part.local_count(0) == 2, "large block comes first");
-          BULK_CHECK(part.local_count(1) == 1, "small block comes after");
-          BULK_CHECK(part.local_count(3) == 1, "last block is small");
-          BULK_CHECK(part.owner(0) == 0, "computes owner for 0 correctly")
-          BULK_CHECK(part.owner(1) == 0, "computes owner for 1 correctly")
-          BULK_CHECK(part.owner(3) == 2, "computes owner for 3 correctly")
-          BULK_CHECK(part.owner(4) == 3, "computes owner for 4 correctly")
+            auto part = bulk::block_partitioning<1>(5, 4);
+            BULK_CHECK(part.local_count(0) == 2, "large block comes first");
+            BULK_CHECK(part.local_count(1) == 1, "small block comes after");
+            BULK_CHECK(part.local_count(3) == 1, "last block is small");
+            BULK_CHECK(part.owner(0) == 0, "computes owner for 0 correctly")
+            BULK_CHECK(part.owner(1) == 0, "computes owner for 1 correctly")
+            BULK_CHECK(part.owner(3) == 2, "computes owner for 3 correctly")
+            BULK_CHECK(part.owner(4) == 3, "computes owner for 4 correctly")
+
+            BULK_CHECK(part.local(1) == 1,
+                       "local: first processor has two elements");
+            BULK_CHECK(part.local(2) == 0,
+                       "local: other processors have one element (2)");
+            BULK_CHECK(part.local(3) == 0,
+                       "local: other processors have one element (3)");
+            BULK_CHECK(part.local(4) == 0,
+                       "local: other processors have one element (4)");
+
+            BULK_CHECK(part.global(1, 0) == 1,
+                       "global: first processor has two elements");
+            BULK_CHECK(part.global(0, 1) == 2,
+                       "global: other processors have one element (2)");
+            BULK_CHECK(part.global(0, 2) == 3,
+                       "global: other processors have one element (3)");
+            BULK_CHECK(part.global(0, 3) == 4,
+                       "global: other processors have one element (4)");
         }
-
-
     });
 }